+++ client/src/App.js
@@ -30,8 +30,8 @@ function App() {
         <BrowserRouter>   
             <Routes>
                 <Route exact path='/' element={<Home menu={menu} setMenu={setMenu}/>}/>
-                <Route exact path='/register' element={<Register/>}/>
-                <Route exact path='/login' element={<Login/>}/>
+                <Route exact path='/register' element={<Register menu={menu} setMenu={setMenu}/>}/>
+                <Route exact path='/login' element={<Login menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/attend' element={<Attend menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/wifi-attend' element={<Wifi menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/gps-attend' element={<GPS menu={menu} setMenu={setMenu}/>}/>
@@ -43,7 +43,7 @@ function App() {
                 <Route exact path='/loginhistory' element={<LoginHistory menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/board' element={<Board menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/report' element={<Report menu={menu} setMenu={setMenu}/>}/>
-                <Route exact path='/check-email' element={<CheckEmail/>}/>
+                <Route exact path='/check-email' element={<CheckEmail menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/confirm/:confirmationCode' element={<Confirm menu={menu} setMenu={setMenu}/>}/>
                 <Route exact path='/too-many-requests' element={<TooManyRequests/>}/>
                 <Route path='*' element={<NotFound/>} />