+++ client/src/pages/Register.jsx
@@ -2,10 +2,10 @@ import Navbar from '../components/Navbar'
 import Auth from '../components/Auth'
 import Footer from '../components/Footer'
 
-const Register = () => {   
+const Register = ({menu, setMenu}) => {   
     return (
         <div>
-            <Navbar/> 
+            <Navbar menu={menu} setMenu={setMenu}/> 
             <Auth
                 mode='register'
             />